% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_it_really.R
\name{areYouTaken}
\alias{areYouTaken}
\title{A function that simply asks 'Are you taken?'}
\usage{
areYouTaken(name, status = "S", statusDuration)
}
\arguments{
\item{name}{A string name.}

\item{status}{A character status indicator ("S" = Single, "M" = Married, "D" = Divorced).}

\item{statusDuration}{A numeric value for the number of years this status has been in effect.}
}
\value{
Returns nothing, prints output.
}
\description{
A function that simply asks 'Are you taken?'
}
