% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeYourRStudio.R
\name{closeYourRStudio}
\alias{closeYourRStudio}
\title{A function that has no purpose other than to scold you before quitting your r session.}
\usage{
closeYourRStudio(reason = "")
}
\arguments{
\item{reason}{A string value that is empty by default.}
}
\value{
Nothing, this function returns you to your
}
\description{
Why? more like Why not!
}
\examples{
closeYourRStudio("I no longer yearn for code")
}
