% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myStockPortfolio.R
\name{myStockPortfolio}
\alias{myStockPortfolio}
\title{Shows you the dollar value of a stock portfolio of someone who has never invested their money.}
\usage{
myStockPortfolio()
}
\value{
Nothing, but it does print a plot
}
\description{
Shows you the dollar value of a stock portfolio of someone who has never invested their money.
}
\examples{
myStockPortfolio()
}
